Fail if bin/runner is called with no option flags
The argument checks goes like this:

- If the first argument is `-r`, then pass the second argument to
  `rapidjson_usage_test()`

- If the first argument is `-bp`, then pass the second argument to
  `cjson_usage_test()`

- Else, pass the *second* argument to `cjson_usage_test()`

The usage output says:

```
runner [-rc] <file.json>
```

Which makes me believe that the option flags are indeed optional,
therefore passing a file as a first argument should work, however it
causes a segmentation fault, as `argv[2]` points to nothing.

Instead of making the arguments optional, this commit makes the first
option required.

Signed-off-by: Juan Cruz Viotti <jv@jviotti.com>